% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_bbox.R
\name{expand_bbox}
\alias{expand_bbox}
\title{Function to extract bounding box of object
Useful for clipping basemaps to study region}
\usage{
expand_bbox(shapefile, expand = 0.1, equal = T)
}
\description{
Function to extract bounding box of object
Useful for clipping basemaps to study region
}
