% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_merriweather}
\alias{install_merriweather}
\title{Install the Merriweather Fonts}
\usage{
install_merriweather(message = F, force = F)
}
\description{
This function checks whether the Merriweather font is installed, and if not will
 install the package
}
