% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_style.R
\name{set_fm_style}
\alias{set_fm_style}
\title{Set FM ggplot theme}
\usage{
set_fm_style(...)
}
\description{
This function allows you to set the ggplot theme of your graphics
}
\details{
@param ... additional arguments which are processed by the `theme` argument
  of ggplot
}
\examples{
set_fm_style()
}
\keyword{fm_style}
